delete TemplateResolverProvider, DetailedErrorComparator, admin.html, add  adminController
