+++ .gitignore
@@ -33,3 +33,4 @@ build/
 .vscode/
 #########properties
 application.properties
+/bin/
