+++ src/main/java/ml/kalanblowsystemmanagement/security/form/LoggingAccessDeniedHandler.java
@@ -28,7 +28,7 @@ public class LoggingAccessDeniedHandler implements AccessDeniedHandler {
 					+ request.getRequestURI());
 		}
 
-		response.sendRedirect(request.getContextPath() + "error/403error");
+		response.sendRedirect(request.getContextPath() + "/access_denied");
 
 	}
 }
