--- src/main/java/ml/kalanblowsystemmanagement/utils/TemplateResolverProvider.java
@@ -1,21 +1,0 @@
-/*
- * package ml.kalanblowsystemmanagement.utils;
- * 
- * import org.thymeleaf.templatemode.TemplateMode; import
- * org.thymeleaf.templateresolver.ClassLoaderTemplateResolver;
- * 
- * import lombok.experimental.UtilityClass;
- * 
- * @UtilityClass public class TemplateResolverProvider {
- * 
- * public ClassLoaderTemplateResolver chapterTemplateResolver(String prefix) {
- * ClassLoaderTemplateResolver emailTemplateResolver = new
- * ClassLoaderTemplateResolver(); emailTemplateResolver.setPrefix(prefix);
- * emailTemplateResolver.setSuffix(".html");
- * emailTemplateResolver.setTemplateMode(TemplateMode.HTML);
- * emailTemplateResolver.setCharacterEncoding("UTF-8");
- * emailTemplateResolver.setOrder(0);
- * emailTemplateResolver.setCheckExistence(true);
- * 
- * return emailTemplateResolver; } }
- */
