+++ src/main/resources/i18n/messages.properties
@@ -11,10 +11,10 @@ menu.reports=Rapports
 menu.profile=votre Profil
 menu.settings=Paramètres
 menu.signout=Se déconnecter
-message.username=Username required
+message.username=Nom d'utilisateur requis
 message.password=Mot de passe requis
-message.unauth=Unauthorized Access !!
-message.badCredentials=Invalid Credentials
+message.unauth=Accès non autorisé !
+message.badCredentials=Informations d'identification non valides
 message.sessionExpired=La session a expiré
 message.logoutError=Désolé, erreur de déconnexion
 message.logoutSucc=Vous vous êtes déconnecté avec succès
@@ -26,7 +26,7 @@ email.not.exists= Email n'existe pas.
 message.lastName=Le nom de famille est requis
 message.firstName=Prénom obligatoire
 message.badEmail=Adresse e-mail invalide
-message.email.config.error=Error in java mail configuration
+message.email.config.error=Erreur dans la configuration du courrier java
 PastLocalDate.profileForm.birthDate=Veuillez donner votre vraie date de naissance
 NotNull.profileForm.birthDate=Veuillez spécifier votre date de naissance
 typeMismatch.birthDate = Date de naissance invalide.
@@ -41,7 +41,7 @@ label.user.confirmPass=Confirmez votre mot de passe
 label.user.mobileNumber=Numéro de mobile
 label.user.birthDate=Date de naissance
 label.user.birthDate.palceholder= E.g : 20/05/1998
-label.user.2fa=Use Two step verification
+label.user.2fa=Utilisez la vérification en deux étapes
 label.form.submit=Envoyer
 label.form.title=Fiche d'inscription
 label.form.loginLink=Retour au login
@@ -143,35 +143,35 @@ NotNull.user.firstName=Prénom est requis
 NotEmpty.user.firstName=Prénom est requis
 NotNull.user.lastName=Nom de famille requis
 NotEmpty.user.lastName=Nom de famille requis
-NotNull.user.username=Username(Email) required
-NotEmpty.user.username=Username(Email) required
-NotNull.user.password=Mot de passe requisd
+NotNull.user.username=Nom d'utilisateur(Email) requis
+NotEmpty.user.username=Nom d'utilisateur(Email) requis
+NotNull.user.password=Mot de passe requis
 NotEmpty.user.password=Mot de passe requis
-NotNull.user.matchingPassword=Required
-NotEmpty.user.matchingPassword=Required
-Email.user.email=Invalid Username (Email)
-label.form.resendRegistrationToken=Re-send Token
-message.resendToken=We will send an email with a new registration token to your email account
-message.forgetPassword=Forget Password
-message.resetPassword=Reset Password
-message.updatePassword=Update Password
-message.userNotFound=User Not Found
-auth.message.blocked=This ip is blocked for 24 hours
-message.accountVerified=Your account verified successfully
-message.resetPasswordSuc=Password reset successfully
-message.resetYourPassword=Reset your password
-message.resetPasswordEmail=You should receive an Password Reset Email shortly
-message.error=Error Occurred
-message.updatePasswordSuc=Password updated successfully
-message.changePassword=Change Password
-message.invalidOldPassword=Invalid Old Password
-label.user.newPassword=New Password
-label.user.oldPassword=Old Password
-error.wordLength=Your password is too short
-error.wordNotEmail=Do not use your email as your password
-error.wordSequences=Your password contains sequences
-error.wordLowercase=Use lower case characters
-error.wordUppercase=Use upper case characters
+NotNull.user.matchingPassword=Requis
+NotEmpty.user.matchingPassword=Requis
+Email.user.email=Nom d'utilisateur (Email) non valide
+label.form.resendRegistrationToken=Renvoi du jeton
+message.resendToken=Nous enverrons un courriel avec un nouveau jeton d'enregistrement à votre compte de messagerie.
+message.forgetPassword=Mot de passe oublié
+message.resetPassword=Réinitialiser le mot de passe
+message.updatePassword=mettre à jour le mot de passe
+message.userNotFound=Utilisateur non trouvé
+auth.message.blocked=Cette IP est bloquée pendant 24 heures
+message.accountVerified=Votre compte a été vérifié avec succès
+message.resetPasswordSuc=Réinitialisation du mot de passe avec succès
+message.resetYourPassword=Réinitialiser votre mot de passe
+message.resetPasswordEmail=Vous devriez recevoir un courriel de réinitialisation du mot de passe sous peu
+message.error=Une erreur s'est produite
+message.updatePasswordSuc=Le mot de passe a été mis à jour avec succès
+message.changePassword=Changer le mot de passe
+message.invalidOldPassword=Ancien mot de passe invalide
+label.user.newPassword=Nouveau mot de passe
+label.user.oldPassword=Ancien mot de pass
+error.wordLength=Votre mot de passe est trop court
+error.wordNotEmail=N'utilisez pas votre adresse électronique comme mot de passe
+error.wordSequences=Votre mot de passe contient des séquences
+error.wordLowercase=Utilisez des caractères minuscules
+error.wordUppercase=Utilisez des caractères majuscule
 #######Login page##################
 login.page.connexion=Se connecter
 login.page.title=Créer un compte
@@ -198,7 +198,7 @@ user.auth.error=un login ou un mot de passe erroné
 user.del=Supprimer
 user.edit=éditer
 user.name=Nom
-user.roles=role
+user.roles=Role
 user.roles.create= Ajouter un nouveau role
 user.roles.edite=Mettre à jour le role
 users.actions=action des utilisateurss
@@ -230,5 +230,5 @@ NotBlank.user.phoneNumber=Veuillez saisir le numéro de téléphone.
 Pattern.user.phoneNumber=Veuillez saisir un numéro de téléphone valide.
 
 ########################Student Page###############################################
-student.page.label.title=Rgister New Student
+student.page.label.title=Inscription d'un nouvel élève
