+++ src/main/resources/i18n/messages_en.properties
@@ -129,7 +129,7 @@ user.auth.error=wrong login or password
 user.del=remove user
 user.edit=Edit user
 user.name=name
-user.roles=role
+user.roles=Authority
 user.roles.create=Add new Role
 user.roles.edite=Edite Role
 users.actions=users action
