+++ src/main/resources/templates/admin/homepage.html
@@ -1,5 +1,8 @@
 <!DOCTYPE html>
-<html lang="fr" xmlns:th="http://www.thymeleaf.org">
+<html lang="fr" xmlns:th="http://www.thymeleaf.org"
+        xmlns:sec="http://www.thymeleaf.org/extras/spring-security"
+        xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+        layout:decorate="~{layout/layout}">
 <head th:replace="fragments/header :: header">
 
 <title layout:title-pattern="$LAYOUT_TITLE - $CONTENT_TITLE">Kalanblow
