--- src/main/resources/templates/admin/user/admin.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="fr" 
-	xmlns:th="http://www.thymeleaf.org"
-	xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-	layout:decorate="~{layout/layout}">
-<head th:replace="fragments/header :: header">
-<title>Admin</title>
-</head>
-
-<body >
-	
-	<div class="container-fluid" layout:fragment="page-content">
-		<!-- ============================================================== -->
-		<!-- Bread crumb and right sidebar toggle -->
-		<!-- ============================================================== -->
-		<h1 class="h1">Bienvenue [[${#httpServletRequest.remoteUser}]]!</h1>
-		<h2 class="h2">Vous êtes connecté avec succès en tant qu'Admin !</h2>
-		<div sec:authorize="isAuthenticated()">N1 : Ce contenu n'est
-			montré qu'aux utilisateurs authentifiés.</div>
-		<div sec:authorize="hasAnyAuthority('ADMIN')">N2 : Ce contenu
-			n'est montré qu'aux administrateurs.</div>
-		<p>
-			Cliquez sur<a href="/users" class="">Ici</a>pour montrer tous les
-			enregistrements utilisateurs.
-		</p>
-		<form action="/logout" method="post">
-			<input type="submit" value="Sign Out"
-				class="input-group .form-control" />
-		</form>
-	</div>
-</body>
-</html>
