+++ src/main/resources/templates/fragments/footer.html
@@ -2,7 +2,7 @@
 <html lang="fr" xmlns="http://www.w3.org/1999/xhtml"
 	xmlns:th="http://www.thymeleaf.org"
 xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
-<head th:replace="layout/header :: header">
+<head th:replace="fragments/header :: header">
 </head>
 
 <body>
