+++ src/main/resources/templates/fragments/navigation.html
@@ -79,7 +79,7 @@ xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
 					<li class="nav-item d-flex"><i
 						class="align-self-center fas fa-user-cog" style="color: white;"></i>
 						<a class="nav-link" sec:authorize="hasRole('ADMIN')"
-						th:href="@{/adminHome}"> Admin Page </a></li>
+						th:href="@{/adminHome}"> Admin home page </a></li>
 					<!--================================================================================================-->
 					<!--endregion-->
 				</ul>
