+++ src/main/resources/templates/users/userDelete.html
@@ -49,11 +49,11 @@
 			</tr>
 			<tr>
 				<td><a href="#"
-					data-th-href="@{/adminPage/deleteUser(id=${userId.id},phase=confirm)}">
+					data-th-href="@{/users/deleteUser(id=${userId.id},phase=confirm)}">
 						<button type="button" data-th-text="#{delete.button.label}">Delete</button>
 				</a></td>
 				<td><a href="#"
-					data-th-href="@{/adminPage/deleteUser(id=${userId.id},phase=cancel)}">
+					data-th-href="@{/users/deleteUser(id=${userId.id},phase=cancel)}">
 						<button type="button" data-th-text="#{cancel.button.label}">Cancel</button>
 				</a></td>
 			</tr>
